signup: refactor folders structure
