+++ packages/pilot/src/containers/Account/SignUp/index.js
@@ -1,3 +1,4 @@
 export { default as InvalidEmailError } from './InvalidEmailError'
-export { default as SignUpForm } from './Form'
+export { default as SignUpForm } from './LoginSignUp/Form'
 export { default as SignUpConfirmation } from './Confirmation'
+export { default as SignUpInviteForm } from './Invite/Form'
